requirements.txt
appnope==0.1.2
backcall==0.2.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.0
en-core-web-sm==3.2.0
entrypoints==0.3
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
langcodes==3.3.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
nest-asyncio==1.5.4
nltk==3.6.5
numpy==1.21.4
packaging==21.3
parso==0.8.3
pathy==0.6.1
pdfminer.six==20211012
pexpect==4.8.0
pickleshare==0.7.5
preshed==3.0.6
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
PyMuPDF==1.19.2
pyparsing==3.0.6
pyTelegramBotAPI==4.2.2
python-dateutil==2.8.2
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typer==0.4.0
typing-extensions==4.0.1
urllib3==1.26.7
wasabi==0.9.0
wcwidth==0.2.5
